+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add settings option to enable trace logs
 - Add contract details to dlc channels api
 - Add support for parsing invoices from Zeus
+- Add social media links to app info
 
 ## [1.6.4] - 2023-11-24